+++ main.py
@@ -1,0 +1,10 @@
+"""
+This file will run all of the components
+
+1) Data import
+2) Data pre-processing and filtering 
+3) Data formatting 
+4) Feature extraction
+5) Feature selection 
+6) Model training and testing
+"""
